Support IO register addresses >= 256

+++ simavr/sim/sim_regbit.h
@@ -42,7 +42,7 @@ extern "C" {
  */
 static inline uint8_t avr_regbit_set(avr_t * avr, avr_regbit_t rb)
 {
-	uint8_t a = rb.reg;
+	uint16_t a = rb.reg;
 	uint8_t m;
 
 	if (!a)
@@ -54,7 +54,7 @@ static inline uint8_t avr_regbit_set(avr_t * avr, avr_regbit_t rb)
 
 static inline uint8_t avr_regbit_setto(avr_t * avr, avr_regbit_t rb, uint8_t v)
 {
-	uint8_t a = rb.reg;
+	uint16_t a = rb.reg;
 	uint8_t m;
 
 	if (!a)
@@ -69,7 +69,7 @@ static inline uint8_t avr_regbit_setto(avr_t * avr, avr_regbit_t rb, uint8_t v)
  */
 static inline uint8_t avr_regbit_setto_raw(avr_t * avr, avr_regbit_t rb, uint8_t v)
 {
-	uint8_t a = rb.reg;
+	uint16_t a = rb.reg;
 	uint8_t m;
 
 	if (!a)
@@ -81,7 +81,7 @@ static inline uint8_t avr_regbit_setto_raw(avr_t * avr, avr_regbit_t rb, uint8_t
 
 static inline uint8_t avr_regbit_get(avr_t * avr, avr_regbit_t rb)
 {
-	uint8_t a = rb.reg;
+	uint16_t a = rb.reg;
 	if (!a)
 		return 0;
 	//uint8_t m = rb.mask << rb.bit;
@@ -95,7 +95,7 @@ static inline uint8_t avr_regbit_get(avr_t * avr, avr_regbit_t rb)
  */
 static inline uint8_t avr_regbit_from_value(avr_t * avr, avr_regbit_t rb, uint8_t value)
 {
-	uint8_t a = rb.reg;
+	uint16_t a = rb.reg;
 	if (!a)
 		return 0;
 	return (value >> rb.bit) & rb.mask;
@@ -106,7 +106,7 @@ static inline uint8_t avr_regbit_from_value(avr_t * avr, avr_regbit_t rb, uint8_
  */
 static inline uint8_t avr_regbit_get_raw(avr_t * avr, avr_regbit_t rb)
 {
-	uint8_t a = rb.reg;
+	uint16_t a = rb.reg;
 	if (!a)
 		return 0;
 	//uint8_t m = rb.mask << rb.bit;
@@ -115,7 +115,7 @@ static inline uint8_t avr_regbit_get_raw(avr_t * avr, avr_regbit_t rb)
 
 static inline uint8_t avr_regbit_clear(avr_t * avr, avr_regbit_t rb)
 {
-	uint8_t a = (rb.reg);
+	uint16_t a = rb.reg;
 	uint8_t m = rb.mask << rb.bit;
 	avr_core_watch_write(avr, a, avr->data[a] & ~m);
 	return avr->data[a];
@@ -133,7 +133,7 @@ static inline uint8_t avr_regbit_get_array(avr_t * avr, avr_regbit_t *rb, int co
 	int i;
 
 	for (i = 0; i < count; i++, rb++) if (rb->reg) {
-		uint8_t a = (rb->reg);
+		uint16_t a = rb->reg;
 		res |= ((avr->data[a] >> rb->bit) & rb->mask) << i;
 	}
 	return res;
